requirements.txt
cachetools==6.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
db-dtypes==1.4.4
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
influxdb3-python==0.16.0
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandas-gbq==0.30.0
polars==1.35.1
polars-runtime-32==1.35.1
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reactivex==4.0.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
